tests compiled into one
