+++ gulpfile.js
@@ -48,7 +48,7 @@ gulp.task('jscs', function() {
 });
 
 gulp.task('lint', function () {
-  return gulp.src(['app/**/*.js', '!node_modules/**'])
+  return gulp.src(['**/*.js', '!node_modules/**'])
   .pipe(eslint.format())
   .pipe(eslint.failAfterError());
 });
