+++ lib/git_util.js
@@ -79,9 +79,7 @@ function getReadme(req, cb) {
       if (err) console.log(err);
       var payload = JSON.parse(data.text);
       req.fileContent = payload.content;
-      //console.log(payload);
       var contents = new Buffer(payload.content, 'base64').toString();
-      //console.log(contents);
       cb(contents);
     });
 }
