+++ package.json
@@ -48,7 +48,7 @@
   },
   "dependencies": {
     "body-parser": "^1.14.2",
-    "cadet-parser": "^0.2.1",
+    "cadet-parser": "^0.3.1",
     "eat": "^0.1.1",
     "express": "^4.13.3",
     "isomorphic-fetch": "^2.2.0",
